refactor(dir): mark filesystem driver private
